sovled no-saved-game-to-restore issue

+++ cleesh/app_main/file_io.py
@@ -5,6 +5,7 @@
 
 ### import statements
 import shutil
+from pathlib import Path
 
 
 ### main routine
@@ -12,10 +13,16 @@ def save_game(game_name):
     src = f"/Users/tas/Documents/Python/dark_castle3/cleesh/games/{game_name}/working/active_pkl"
     dst = f"/Users/tas/Documents/Python/dark_castle3/cleesh/games/{game_name}/saves/save_pkl"
     shutil.copyfile(src, dst)
-    return
+    user_output = "Game saved."
+    return user_output
 
 def restore_game(game_name):
     src = f"/Users/tas/Documents/Python/dark_castle3/cleesh/games/{game_name}/saves/save_pkl"
-    dst = f"/Users/tas/Documents/Python/dark_castle3/cleesh/games/{game_name}/working/active_pkl"
-    shutil.copyfile(src, dst)
-    return
+    my_file = Path(src)
+    if my_file.is_file():
+        dst = f"/Users/tas/Documents/Python/dark_castle3/cleesh/games/{game_name}/working/active_pkl"
+        shutil.copyfile(src, dst)
+        user_output = "Game restored."
+    else:
+        user_output = "There is no saved game to restore."
+    return user_output
